feat: extract shared game components, add sound effects
Implement the 6 PRD game components that were previously inlined, plus the deferred sound system, and fix pre-existing broken tests.

Components (resources/js/Components/Game/):
- GameLayout, TimerBar, CountdownOverlay, ScoreAnimation, StreakBadge, GameCodeInput — extracted from inline markup in Host/Player pages
- Refactor Host/Game.vue, Player/Game.vue, Player/Join.vue to use them

Sound:
- useSound.js composable — WebAudio-generated tones (no asset files): tick/go/correct/wrong/whoosh/fanfare
- Wired into Host + Player game pages with persisted mute toggle
- Add play.mute/unmute, host.mute/unmute i18n keys (id + en)

Fixes (pre-existing failures):
- AppSetting::all() was an incompatible override of Model::all(), causing a fatal error that crashed the test run; rename to getMany()
- FullGameFlowTest answer payload missing required time_taken field
- ScoringService avg_time cast to float for driver-independent results

Tests: 188 passed, 7 skipped.

+++ app/Services/ScoringService.php
@@ -84,7 +84,7 @@ class ScoringService
                 'score' => $player->score,
                 'correct_count' => $correctCount,
                 'best_streak' => $bestStreak,
-                'avg_time' => $answers->whereNotNull('answer_id')->avg('time_taken') ?? 0,
+                'avg_time' => (float) ($answers->whereNotNull('answer_id')->avg('time_taken') ?? 0),
             ];
         });
     }
@@ -108,7 +108,7 @@ class ScoringService
                     'correct_answers' => $answers->where('is_correct', true)->count(),
                     'total_answers' => $answers->count(),
                     'longest_streak' => $this->computeBestStreak($answers),
-                    'avg_time' => $answers->whereNotNull('answer_id')->avg('time_taken') ?? 0,
+                    'avg_time' => (float) ($answers->whereNotNull('answer_id')->avg('time_taken') ?? 0),
                 ];
             })
             ->keyBy('player_id');
